Implement GLFW event handling
GLFW calls one of the defined handlers which convert this into high
level Event objects and then add this into global event queue. This
queue is global, as we don't define any custom context which would store
this queue for each Window.

Anyway, this is sufficient for now as I don't plan to support multiple
windows here.