Implement basic template for reading .stl files
Here, we create a few class, that are the base for implementing proper
reading .stl files. As we know, .stl files has two type: text and
binary. Each are read differently, so we have to check which file is
this. For this, I created the ReaderFactory class, that reads first five
bytes and check the type. Then return proper class, which is of Reader
interface.